services/remnawave_service/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
remnawave
debugpy
python-dateutil
asyncio
sentry-sdk[fastapi]